[FE] fix : 모임 생성시 카테고리 안되던 오류 fix 및 모달 밖 영역 선택시오류 수정
