+++ frontend/moim_front/src/components/CreationModal/index.jsx
@@ -1,13 +1,11 @@
 import React, { useRef, useEffect } from "react";
 import { postMeeting } from "../../api/moim";
 import { parse } from "date-fns";
-import axios from "axios";
-
+import { POST } from "../../utils/axios";
 function CreationModal({
   showModal,
   setShowModal,
   children,
-  closeHandler,
   isMeeting,
   moimId,
   startDateTime,
@@ -20,46 +18,39 @@ function CreationModal({
 }) {
   const modalRef = useRef();
 
-  if (isMeeting) {
-  }
   useEffect(() => {
     console.log(startDateTime);
     console.log(endDateTime);
   }, []);
-  const closeModal = () => {
-    closeHandler();
-    setShowModal(false);
 
-    if (isMeeting) {
-      const data = {
-        moimId: parseInt(moimId),
-        agenda: agenda,
-        startDateTime: startDateTime.replace("T", " ").split("+")[0],
-        endDateTime: endDateTime.replace("T", " ").split("+")[0],
-        place: place,
-        meetingCategory: "REGULAR",
-      };
-      postMeeting(data);
+  const createMeeting = async () => {
+    const data = {
+      moimId: parseInt(moimId),
+      agenda: agenda,
+      startDateTime: startDateTime.replace("T", " ").split("+")[0],
+      endDateTime: endDateTime.replace("T", " ").split("+")[0],
+      place: place,
+      meetingCategory: meetingCategory,
+    };
+    console.log("create Meeting");
+    try {
+      const response = await POST("api/meetings", data);
+      console.log(response);
       setOpen(true);
-    } else {
-      //Personal 스케쥴 등록
-      console.log(endDateTime);
-      // const data = {
-      //   scheduleName: scheduleTitle,
-      //   startDateTime: startDateTime.replace("T", " ").split("+")[0],
-      //   endDateTime: endDateTime.replace("T", " ").split("+")[0],
-      // };
-      // try {
-      //   const result = async () => await axios.post("api/schedules", data);
-      //   console.log(result);
-      // } catch (e) {
-      //   console.log(e);
-      // }
+    } catch (e) {
+      console.log(e);
+    }
+  };
+
+  const closeModal = () => {
+    setShowModal(false);
+    if (isMeeting) {
+      createMeeting();
     }
   };
 
   const modalOutSideClick = (e) => {
-    if (modalRef.current === e.target) closeModal();
+    if (modalRef.current === e.target) setShowModal(false);
   };
 
   return (
